requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
dill==0.3.6
discord.py==2.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
multidict==6.0.2
nodeenv==1.7.0
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
platformdirs==2.5.4
pycodestyle==2.9.1
pylint==2.15.6
pyright==1.1.280
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
yarl==1.8.1
